Revert "[LLVM] Add flags to crash the opt/codegen pipeline (#200967)"
This reverts commit 015d0479b94ce48e431398813525a59a16c6e4eb.
