+++ llvm/utils/gn/secondary/llvm/lib/Transforms/Utils/BUILD.gn
@@ -99,7 +99,6 @@ static_library("Utils") {
     "StripGCRelocates.cpp",
     "StripNonLineTableDebugInfo.cpp",
     "SymbolRewriter.cpp",
-    "TriggerCrashPass.cpp",
     "UnifyFunctionExitNodes.cpp",
     "UnifyLoopExits.cpp",
     "Utils.cpp",
